+++ tests/test_data_source_local.py
@@ -73,16 +73,67 @@ def test_find_events_by_id_from_data_provider(demo_data_source: DataSource):
     for event_ in events:
         event_["attachedMessageIds"].sort()
 
+    broken_event: dict = data_source.find_events_by_id_from_data_provider("id", True)
+    broken_events: list = data_source.find_events_by_id_from_data_provider(["id", "ids"], True)
+    plug_for_broken_vent: dict =  {
+            "attachedMessageIds": {},
+            "bathcId": 'Broken_Event',
+            "endTimestamp": {"nanao": 0, "epochSecond": 0},
+            "startTimestamp": {"nanao": 0, "epochSecond": 0},
+            "type": 'event',
+            'eventID': 'id',
+            "eventName":'Broken_Event',
+            "eventType":'Broken_Event',
+            "parentEventId": 'Broken_Event',
+            "successful": None,
+            "isBatched": None
+            }
+    plug_for_broken_vents: list = [{
+            "attachedMessageIds": {},
+            "bathcId": 'Broken_Event',
+            "endTimestamp": {"nanao": 0, "epochSecond": 0},
+            "startTimestamp": {"nanao": 0, "epochSecond": 0},
+            "type": 'event',
+            'eventID': 'id',
+            "eventName":'Broken_Event',
+            "eventType":'Broken_Event',
+            "parentEventId": 'Broken_Event',
+            "successful": None,
+            "isBatched": None
+            }, {
+            "attachedMessageIds": {},
+            "bathcId": 'Broken_Event',
+            "endTimestamp": {"nanao": 0, "epochSecond": 0},
+            "startTimestamp": {"nanao": 0, "epochSecond": 0},
+            "type": 'event',
+            'eventID': 'ids',
+            "eventName":'Broken_Event',
+            "eventType":'Broken_Event',
+            "parentEventId": 'Broken_Event',
+            "successful": None,
+            "isBatched": None
+            }]
+
     # Check types
     assert isinstance(event, dict)
     assert isinstance(events, list)
     assert isinstance(events_with_one_element, list)
+    assert isinstance(broken_event, dict)
+    assert isinstance(broken_events, list)
     # Check content.
     assert event == expected_event
     assert events == expected_events
     assert len(events) == 2
     assert len(events_with_one_element) == 1
-
+    # Check Broken_Events
+    assert broken_event == plug_for_broken_vent
+    assert broken_events == plug_for_broken_vents
+    assert [event, broken_event] == data_source.find_events_by_id_from_data_provider(
+        ["88a3ee80-d1b4-11eb-b0fb-199708acc7bc", "id"], True)
+    with pytest.raises(ValueError):
+        data_source.find_events_by_id_from_data_provider(["88a3ee80-d1b4-11eb-b0fb-199708acc7bc", "id"])
+    with pytest.raises(ValueError):
+        data_source.find_events_by_id_from_data_provider("id")
 
 def test_find_messages_by_id_from_data_provider(demo_data_source: DataSource):
     data_source = demo_data_source